+++ lib/collection/src/operations/mod.rs
@@ -16,6 +16,7 @@ pub mod universal_query;
 pub mod validation;
 pub mod vector_ops;
 pub mod vector_params_builder;
+pub mod verification;
 
 use std::collections::HashMap;
