+++ src/actix/api/search_api.rs
@@ -8,13 +8,14 @@ use collection::operations::types::{
 };
 use futures::TryFutureExt;
 use itertools::Itertools;
+use storage::content_manager::collection_verification::CollectionRequestVerification;
 use storage::dispatcher::Dispatcher;
 use tokio::time::Instant;
 
 use super::read_params::ReadParams;
 use super::CollectionPath;
 use crate::actix::auth::ActixAccess;
-use crate::actix::helpers::{self, process_response};
+use crate::actix::helpers::{self, process_response, process_response_error};
 use crate::common::points::{
     do_core_search_points, do_search_batch_points, do_search_point_groups, do_search_points_matrix,
 };
@@ -27,6 +28,11 @@ async fn search_points(
     params: Query<ReadParams>,
     ActixAccess(access): ActixAccess,
 ) -> HttpResponse {
+    let pass = match request.check(&dispatcher, &access, &collection.name).await {
+        Ok(pass) => pass,
+        Err(err) => return process_response_error(err, Instant::now()),
+    };
+
     let SearchRequest {
         search_request,
         shard_key,
@@ -39,7 +45,7 @@ async fn search_points(
 
     helpers::time(
         do_core_search_points(
-            dispatcher.toc(&access),
+            dispatcher.toc_new(&access, &pass),
             &collection.name,
             search_request.into(),
             params.consistency,
@@ -65,6 +71,11 @@ async fn batch_search_points(
     params: Query<ReadParams>,
     ActixAccess(access): ActixAccess,
 ) -> HttpResponse {
+    let pass = match request.check(&dispatcher, &access, &collection.name).await {
+        Ok(pass) => pass,
+        Err(err) => return process_response_error(err, Instant::now()),
+    };
+
     let request = request.into_inner();
     let requests = request
         .searches
@@ -86,7 +97,7 @@ async fn batch_search_points(
 
     helpers::time(
         do_search_batch_points(
-            dispatcher.toc(&access),
+            dispatcher.toc_new(&access, &pass),
             &collection.name,
             requests,
             params.consistency,
