+++ src/actix/mod.rs
@@ -5,7 +5,6 @@ mod auth;
 mod certificate_helpers;
 #[allow(dead_code)] // May contain functions used in different binaries. Not actually dead
 pub mod helpers;
-pub mod request_verification;
 pub mod web_ui;
 
 use std::io;
